requirements.txt
# 崖山数据库 MCP Server 依赖
# ================================

# MCP 核心框架
mcp>=1.0.0

# JDBC 桥接驱动（崖山数据库必须使用 JDBC）
jpype1>=1.4.0
jaydebeapi>=1.2.0

# Web 服务器（用于 HTTP/SSE 模式）
uvicorn>=0.24.0
fastapi>=0.104.0
sse-starlette>=2.0.0
